tasklist-kadai

+++ app/Http/Controllers/TasksController.php
@@ -11,13 +11,21 @@ class TasksController extends Controller
     
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+        //　認証済みユーザを取得
+        $user = \Auth::user();
+        // ユーザーの投稿の一覧を作成日時の降順で取得
+        $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
         
-        return view('tasks.index', [
+        $data = [
             'tasks' => $tasks,
-            ]);
+        ];
     }
-
+    
+    // welcomeビューでそれらを表示
+    return view('welcome', $data);
+    }
     
     public function create()
     {
@@ -33,14 +41,15 @@ class TasksController extends Controller
     {
         // バリデーション
         $request->validate([
-            'status' => 'required|max:10',   // 追加
+            'status' => 'required|max:10',   
             'content' => 'required|max:255',
         ]);
         
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+            ]);
         
         return redirect('/');
     }
@@ -50,9 +59,13 @@ class TasksController extends Controller
     {
         $task = Task::findOrFail($id);
         
-        return view('tasks.show', [
+        // 認証済みユーザー（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if(\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
             'task' => $task,
             ]);
+        }
+        return redirect('/');
     }
 
     
@@ -60,34 +73,45 @@ class TasksController extends Controller
     {
         $task = Task::findOrFail($id);
         
+        if(\Auth::id() === $task->user_id) {
         return view('tasks.edit', [
             'task' => $task,
             ]);
+        }
+            return redirect('/');
     }
 
     
     public function update(Request $request, $id)
     {
+        // バリデーション
         $request->validate([
             'status' => 'required|max:10',   // 追加
             'content' => 'required|max:255',
         ]);
         
+        // メッセージを作成
         $task = Task::findOrFail($id);
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
         
+        // トップページにレダイレクトさせる
         return redirect('/');
     }
 
     
     public function destroy($id)
     {
-        $task = Task::findOrFail($id);
+        // idの値で投稿を検索して取得
+        $task = \App\Task::findOrFail($id);
         
-        $task->delete();
+        // 認証済みユーザー（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if(\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
         
+        // 前のURLへリダイレクトさせる
         return redirect('/');
     }
 }
